+++ src/components/NewPost/style.css
@@ -15,7 +15,18 @@
 
 .MuiFormControl-root .MuiTextField-root{ 
     margin: 10px;
+}  
+
+.post-submitted-div{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
 } 
+
+.post-submitted-text{
+    color: green;
+}
 
 .checkbox-section{ 
     margin: 5px;
