+++ src/firebase/firebase.utils.js
@@ -7,7 +7,7 @@ const config = firebaseConfig;
 
 export const createUserProfileDocument = async(userAuth, additionalData) => {
     if(!userAuth) return;
-    const userRef = firestore.doc(`users/${userAuth.uid}`);
+    const userRef = database.doc(`users/${userAuth.uid}`);
     const snapShot = await userRef.get();
 
 
@@ -33,5 +33,5 @@ export const createUserProfileDocument = async(userAuth, additionalData) => {
 firebase.initializeApp(config);
 
 export const auth = firebase.auth();
-export const firestore = firebase.firestore();
+export const database = firebase.firestore();
